--- src/components/legal_disclaimer.js
@@ -1,12 +1,0 @@
-const legalDisclaimer = [
-  {
-    name: 'Freelances Lyonnais',
-    headquarters: '18-20 rue Tronchet 69006 Lyon, Immatriculée au RCS de LYON : N° 880 240 130 00018,Numéro individuel d\'identification fiscale : FR65880240130',
-    mail: 'Pierre Ammeloot,pierre@webanyone.fr',
-    phone: '+33 6 02 12 06 33',
-    info: 'Hébergeur du site internet : OVH, 2 rue Kellermann, 59100 Roubaix, France',
-    editor: 'Pierre Ammeloot'
-  }
-];
-
-export default legalDisclaimer;
