+++ src/pages/Listing.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Freelance from '../components/Freelance2';
 import '../styles/Listing.scss';
-import freelances from './people';
+import freelances from '../test/people';
 import { Link } from 'react-router-dom';
 
 function Listing () {
