--- src/pages/SearchResult.js
@@ -1,17 +1,0 @@
-import React from 'react';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
-
-function SearchResult () {
-  return (
-    <div>
-      <Header />
-
-      I am SearchResult !
-
-      <Footer />
-    </div>
-  );
-}
-
-export default SearchResult;
