+++ src/test/people.js
@@ -4,7 +4,7 @@ const freelances = [
     siret: '1234567890123',
     firstname: 'John',
     lastname: 'Doe',
-    urlPhoto: require('../images/photos/01.jpg'),
+    urlPhoto: require('./photos/01.jpg'),
     phone: '+33 6 00 00 00 00',
     bio: 'Honores quantum enim possit etiam Non si perducere diligas etiam possis quamvis P videndum consulem amplissimos fratrem si fratrem potuit tribuendum honores non L Non P ut efficere etiam autem sustinere tuos Scipio potuit ad tamen quantum adiuves omnes quantum Rupilium efficere deinde perducere amplissimos quantum videndum Quod deferre adiuves opt.',
     isActive: true,
@@ -15,7 +15,7 @@ const freelances = [
     siret: '12345678901325',
     firstname: 'Martin',
     lastname: 'Saville',
-    urlPhoto: require('../images/photos/02.jpg'),
+    urlPhoto: require('./photos/02.jpg'),
     phone: '04.81.41.24.70',
     bio: 'minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit',
     isActive: true,
@@ -26,7 +26,7 @@ const freelances = [
     siret: '9876543210987',
     firstname: 'Caya',
     lastname: 'Ganelon',
-    urlPhoto: require('../images/photos/03.jpg'),
+    urlPhoto: require('./photos/03.jpg'),
     phone: '01.30.70.17.62',
     bio: 'in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia',
     isActive: true,
@@ -37,7 +37,7 @@ const freelances = [
     siret: '5824673190546',
     firstname: 'Roland',
     lastname: 'Lanteigne',
-    urlPhoto: require('../images/photos/04.jpg'),
+    urlPhoto: require('./photos/04.jpg'),
     phone: '03.02.16.17.25',
     bio: 'Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa',
     isActive: true,
@@ -48,7 +48,7 @@ const freelances = [
     siret: '3574159820486',
     firstname: 'Angelette',
     lastname: 'Baron',
-    urlPhoto: require('../images/photos/05.jpg'),
+    urlPhoto: require('./photos/05.jpg'),
     phone: '02.11.21.50.57',
     bio: 'quae ab illo inventore veritatis et quasi architecto beatae vitae dicta sunt explicabo. Nemo enim ipsam voluptatem quia voluptas sit',
     isActive: true,
